requirements.txt
appnope==0.1.0
Babel==2.3.4
boto3==1.4.4
botocore==1.5.38
colorama==0.3.7
decorator==4.0.10
dj-database-url==0.4.2
Django==1.9.12
django-appconf==1.0.2
django-compressor==2.1
django-debug-toolbar==1.6
django-extensions==1.7.5
django-extra-views==0.6.4
django-haystack==2.5.1
django-oscar==1.3
django-storages==1.5.2
django-tables2==1.0.7
django-treebeard==4.1.0
django-widget-tweaks==1.4.1
docutils==0.13.1
factory-boy==2.8.1
Faker==0.7.7
flake8==3.2.1
flake8-blind-except==0.1.1
flake8-debugger==1.4.0
gunicorn==19.7.1
ipdb==0.10.1
ipython==5.1.0
ipython-genutils==0.1.0
isort==4.2.5
jmespath==0.9.2
mccabe==0.5.3
mock==2.0.0
mysqlclient==1.3.9
olefile==0.44
pbr==1.10.0
pexpect==4.2.1
phonenumbers==7.7.5
pickleshare==0.7.4
Pillow==3.4.2
prompt-toolkit==1.0.9
psycopg2==2.6.2
ptyprocess==0.5.1
purl==1.3
pycodestyle==2.2.0
pycountry==16.11.27.1
pyflakes==1.3.0
Pygments==2.1.3
PyMySQL==0.6.2
pyprof2calltree==1.4.0
pysolr==3.6.0
python-dateutil==2.6.0
pytz==2016.10
rcssmin==1.0.6
requests==2.12.4
rjsmin==1.0.12
s3transfer==0.1.10
simplegeneric==0.8.1
six==1.10.0
sorl-thumbnail==12.4a1
sqlparse==0.2.2
traitlets==4.3.1
Unidecode==0.4.19
wcwidth==0.1.7
whitenoise==3.2.2
Whoosh==2.7.4
win-unicode-console==0.5
